Bulk of Remaining fixes (Errors may still persist but this knocks a lot of them out
Hopefully my final update to weapon systems.   I have touched 90+ % of weapons, Ammo and Ammo Boxes.

A LOT of  Weapons had data that was... incomplete due to ***APPARENT*** rounding... eg Tonnage of 0 when it should be 0.5 or 2 instead of 1.5!  I don't know how all these errors entered the picture but they were there.